refactor: profiler implementation (#1775)
The purpose of this PR is to refactor numeric and text column profiler implementation. The new approach has more modular and easy-to-maintain code. Lots of nested if/else and for loop logic have been refactored.

Apart from unit tests, to make sure that there is nothing changed in the cloud, I also made e2e tests, and the screenshots are attached.