fix: confidence bounds of anomaly detection sit too close together (#1780)
* This PR resolves: anomaly score sends alerts when actual value is VERY close to predicted value

This PR implements the following:

Instead of rounding the uncertainty intervals' lower and upper bounds to ten decimals, they are rounded using np.floor and np.ceil respectively if the metric that we're classifying is an integer. Running the original time series that was causing issues now no longer causes any issues as can be seen on dev. The original check measurements that were used to create that plot can be found through datadog
Notes

This is not the final fix. Inspecting the response that was sent by the backend for this dataset shows that the uncertainty bounds are too close together. This behaviour is caused by the fact that we use the new measurement (the value that we are classifying) to fit the model. This means that the estimates of the prediction errors are way too conservative. I.e. doing something like

m = Prophet()
m.fit(time_series)
predictions = m.predict(time_series)
means that the prediction intervals are extremely narrow and sit unrealistically close to the real_data column. If we would actually use an actual forecast, i.e. something like

m = Prophet()
m.fit(time_series[:-1])
future = m.make_future_dataframe(periods=1)
forecast = m.predict(future)
the prediction intervals around the training data used to fit the model are still very small, but the one used for the prediction are much wider (and would look a lot more realistic).

* hotfix for dask test profile columns (#1786)

---------

Co-authored-by: Vijay Kiran <mail@vijaykiran.com>
Co-authored-by: Baturay Ofluoglu <baturayofluoglu@gmail.com>