Add PR template for soda-extensions reminder
We've had a few times where the soda-extensions package was broken as a
result of changes in soda-core. There is a workflow in place to validate
they work together correctly, but this is a manual process.

This PR adds a template to promote good PR practices, including a
reminder to run the soda-extensions workflow for validating
compatibility.